requirements.txt
brainflow>=5.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pyserial>=3.5
joblib>=1.3.0
